+++ src/dflow/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for permitify project.
+Django settings for dflow project.
 
 Generated by 'django-admin startproject' using Django 2.0.
 
@@ -55,7 +55,7 @@ MIDDLEWARE = [
     'whitenoise.middleware.WhiteNoiseMiddleware',    
 ]
 
-ROOT_URLCONF = 'permitify.urls'
+ROOT_URLCONF = 'dflow.urls'
 
 TEMPLATES = [
     {
@@ -68,7 +68,7 @@ TEMPLATES = [
                 'django.template.context_processors.request',
                 'django.contrib.auth.context_processors.auth',
                 'django.contrib.messages.context_processors.messages',
-                'permitify.context_processors.export_vars'
+                'dflow.context_processors.export_vars'
             ],
         },
     },
