+++ src/dflow/urls.py
@@ -1,4 +1,4 @@
-"""permitify URL Configuration
+"""dflow URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.0/topics/http/urls/
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.urls import path
 
-from permitify import views
+from dflow import views
 
 urlpatterns = [
     path('', views.index),
